requirements.txt
streamlit
pillow
exif
folium
streamlit-folium
